@@ -6,7 +6,7 @@
   "dependencies": {
     "@hookform/resolvers": "^2.8.8",
     "@testing-library/jest-dom": "^5.16.3",
-    "@testing-library/react": "^12.1.4",
+    "@testing-library/react": "^13.0.0",
     "@testing-library/user-event": "^14.0.4",
     "@types/jest": "^27.4.1",
     "@types/node": "^17.0.23",
@@ -18,9 +18,9 @@
     "bootstrap": "^5.1.3",
     "buffer": "^6.0.3",
     "date-fns": "^2.28.0",
-    "react": "^17.0.2",
+    "react": "^18.0.0",
     "react-bootstrap": "^2.2.2",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.0.0",
     "react-hook-form": "^7.29.0",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.0",
@@ -33,6 +33,7 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
+    "testall": "CI=true react-scripts test",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
